requirements.txt
fastapi
uvicorn
torch
torchvision
joblib
pillow
scikit-learn
numpy
